+++ src/Abstracts/Brickable.php
@@ -8,55 +8,43 @@ use Okipa\LaravelBrickables\Models\Brick;
 
 abstract class Brickable
 {
-    /** @property string $brickModelClass */
-    protected $brickModelClass;
+    protected string $brickModelClass;
 
-    /** @property string $bricksControllerClass */
-    protected $bricksControllerClass;
+    protected string $bricksControllerClass;
 
-    /** @property string $label */
-    protected $label;
+    protected string $formViewPath;
 
-    /** @property string $adminViewPath */
-    protected $adminViewPath;
+    protected string $label;
 
-    /** @property string $templateViewPath */
-    protected $templateViewPath;
+    protected string $brickViewPath;
 
-    /** @property array $storeValidationRules */
-    protected $storeValidationRules;
+    protected array $storeValidationRules;
 
-    /** @property array $updateValidationRules */
-    protected $updateValidationRules;
+    protected array $updateValidationRules;
 
-    /** @property string $storeRouteName */
-    protected $storeRouteName;
+    protected string $storeRouteName;
 
-    /** @property string $editRouteName */
-    protected $editRouteName;
+    protected string $editRouteName;
 
-    /** @property string $updateRouteName */
-    protected $updateRouteName;
+    protected string $updateRouteName;
 
-    /** @property string $destroyRouteName */
-    protected $destroyRouteName;
+    protected string $destroyRouteName;
 
-    /** @property string $moveUpRouteName */
-    protected $moveUpRouteName;
+    protected string $moveUpRouteName;
 
-    /** @property string $moveUpRouteName */
-    protected $moveDownRouteName;
+    protected string $moveDownRouteName;
 
-    /**
-     * Brickable constructor.
-     */
+    protected ?string $cssResourcePath;
+
+    protected ?string $jsResourcePath;
+
     public function __construct()
     {
         $this->brickModelClass = $this->setBrickModelClass();
         $this->bricksControllerClass = $this->setBricksControllerClass();
+        $this->brickViewPath = $this->setBrickViewPath();
         $this->label = $this->setLabel();
-        $this->templateViewPath = $this->setBrickViewPath();
-        $this->adminViewPath = $this->setFormViewPath();
+        $this->formViewPath = $this->setFormViewPath();
         $this->storeValidationRules = $this->setStoreValidationRules();
         $this->updateValidationRules = $this->setUpdateValidationRules();
         $this->storeRouteName = $this->setStoreRouteName();
@@ -65,185 +53,95 @@ abstract class Brickable
         $this->destroyRouteName = $this->setDestroyRouteName();
         $this->moveUpRouteName = $this->setMoveUpRouteName();
         $this->moveDownRouteName = $this->setMoveDownRouteName();
+        $this->cssResourcePath = $this->setCssResourcePath();
+        $this->jsResourcePath = $this->setJsResourcePath();
     }
 
-    /**
-     * Set the brickable-related brick model class.
-     *
-     * @return string
-     */
     protected function setBrickModelClass(): string
     {
         return config('brickables.bricks.model');
     }
 
-    /**
-     * Set the brickable-related controller class.
-     *
-     * @return string
-     */
     protected function setBricksControllerClass(): string
     {
         return config('brickables.bricks.controller');
     }
 
-    /**
-     * Set the brickable template view path.
-     *
-     * @return string
-     */
-    protected function setBrickViewPath(): string
-    {
-        return 'laravel-brickables::brickables.' . Str::snake(class_basename($this), '-') . '.brick';
-    }
-
-    /**
-     * Set the brickable admin view path.
-     *
-     * @return string
-     */
-    protected function setFormViewPath(): string
-    {
-        return 'laravel-brickables::brickables.' . Str::snake(class_basename($this), '-') . '.form';
-    }
-
-    /**
-     * Set the brickable store route name.
-     *
-     * @return string
-     */
     protected function setStoreRouteName(): string
     {
         return 'brick.store';
     }
 
-    /**
-     * Set the brickable edit route name.
-     *
-     * @return string
-     */
     protected function setEditRouteName(): string
     {
         return 'brick.edit';
     }
 
-    /**
-     * Set the brickable update route name.
-     *
-     * @return string
-     */
     protected function setUpdateRouteName(): string
     {
         return 'brick.update';
     }
 
-    /**
-     * Set the brickable destroy route name.
-     *
-     * @return string
-     */
     protected function setDestroyRouteName(): string
     {
         return 'brick.destroy';
     }
 
-    /**
-     * Set the brickable move up route name.
-     *
-     * @return string
-     */
     protected function setMoveUpRouteName(): string
     {
         return 'brick.move.up';
     }
 
-    /**
-     * Set the brickable move down route name.
-     *
-     * @return string
-     */
     protected function setMoveDownRouteName(): string
     {
         return 'brick.move.down';
     }
 
-    /**
-     * Get the brickable store validation rules.
-     *
-     * @return array
-     */
     public function getStoreValidationRules(): array
     {
         return $this->storeValidationRules;
     }
 
-    /**
-     * Set the brickable store validation rules.
-     *
-     * @return array
-     */
     abstract protected function setStoreValidationRules(): array;
 
-    /**
-     * Get the brickable update validation rules.
-     *
-     * @return array
-     */
     public function getUpdateValidationRules(): array
     {
         return $this->updateValidationRules;
     }
 
-    /**
-     * Set the brickable update validation rules.
-     *
-     * @return array
-     */
     abstract protected function setUpdateValidationRules(): array;
 
-    /**
-     * Get the management view path.
-     *
-     * @return string
-     */
     public function getFormViewPath(): string
     {
-        return $this->adminViewPath;
+        return $this->formViewPath;
     }
 
-    /**
-     * Get the template view path.
-     *
-     * @return string
-     */
+    protected function setFormViewPath(): string
+    {
+        return 'laravel-brickables::brickables.' . Str::snake(class_basename($this), '-') . '.form';
+    }
+
     public function getBrickViewPath(): string
     {
-        return $this->templateViewPath;
+        return $this->brickViewPath;
     }
 
-    /**
-     * Get the brickable label.
-     *
-     * @return string
-     */
+    protected function setBrickViewPath(): string
+    {
+        return 'laravel-brickables::brickables.' . Str::snake(class_basename($this), '-') . '.brick';
+    }
+
     public function getLabel(): string
     {
         return __($this->label);
     }
 
-    /**
-     * Set the brickable label.
-     *
-     * @return string
-     */
     public function setLabel(): string
     {
         return ucfirst(Str::snake(class_basename($this), ' '));
     }
 
     /**
-     * Get the brickable store route.
-     *
      * @param mixed $parameters
      *
      * @return string
@@ -254,8 +152,6 @@ abstract class Brickable
     }
 
     /**
-     * Get the brickable edit route.
-     *
      * @param mixed $parameters
      *
      * @return string
@@ -266,8 +162,6 @@ abstract class Brickable
     }
 
     /**
-     * Get the brickable update route.
-     *
      * @param mixed $parameters
      *
      * @return string
@@ -278,8 +172,6 @@ abstract class Brickable
     }
 
     /**
-     * Get the brickable destroy route.
-     *
      * @param mixed $parameters
      *
      * @return string
@@ -290,8 +182,6 @@ abstract class Brickable
     }
 
     /**
-     * Get the brickable move up route.
-     *
      * @param mixed $parameters
      *
      * @return string
@@ -302,8 +192,6 @@ abstract class Brickable
     }
 
     /**
-     * Get the brickable move down route.
-     *
      * @param mixed $parameters
      *
      * @return string
@@ -313,23 +201,33 @@ abstract class Brickable
         return route($this->moveDownRouteName, $parameters);
     }
 
-    /**
-     * Get the brickable-related brick model.
-     *
-     * @return \Okipa\LaravelBrickables\Models\Brick
-     */
     public function getBrickModel(): Brick
     {
         return (new $this->brickModelClass);
     }
 
-    /**
-     * Get the brickable-related bricks controller.
-     *
-     * @return \Okipa\LaravelBrickables\Controllers\BricksController
-     */
     public function getBricksController(): BricksController
     {
         return (new $this->bricksControllerClass);
+    }
+
+    public function getCssResourcePath(): ?string
+    {
+        return $this->cssResourcePath;
+    }
+
+    protected function setCssResourcePath(): ?string
+    {
+        return null;
+    }
+
+    public function getJsResourcePath(): ?string
+    {
+        return $this->jsResourcePath;
+    }
+
+    protected function setJsResourcePath(): ?string
+    {
+        return null;
     }
 }