+++ src/Facades/Brickables.php
@@ -6,7 +6,6 @@ use Illuminate\Support\Facades\Facade;
 
 class Brickables extends Facade
 {
-    /** @inheritDoc */
     protected static function getFacadeAccessor()
     {
         return 'Brickables';