+++ tests/Unit/BrickablesTest.php
@@ -3,6 +3,7 @@
 namespace Okipa\LaravelBrickables\Tests\Unit;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Str;
 use Okipa\LaravelBrickables\Abstracts\Brickable;
 use Okipa\LaravelBrickables\Facades\Brickables;
 use Okipa\LaravelBrickables\Models\Brick;
@@ -16,7 +17,7 @@ class BrickablesTest extends BrickableTestCase
     /** @test */
     public function it_can_return_all_registered_brickables()
     {
-        $brickableOne = new Class extends Brickable {
+        $brickableOne = new class extends Brickable {
             protected function setStoreValidationRules(): array
             {
                 return [];
@@ -27,7 +28,7 @@ class BrickablesTest extends BrickableTestCase
                 return [];
             }
         };
-        $brickableTwo = new Class extends Brickable {
+        $brickableTwo = new class extends Brickable {
             protected function setStoreValidationRules(): array
             {
                 return [];
@@ -59,10 +60,10 @@ class BrickablesTest extends BrickableTestCase
     public function it_can_display_model_bricks_html()
     {
         view()->addNamespace('laravel-brickables', 'tests/views');
-        $brickable = new Class extends Brickable {
+        $brickable = new class extends Brickable {
             public function setBrickViewPath(): string
             {
-                return 'laravel-brickables::test-brick';
+                return 'laravel-brickables::brick-test';
             }
 
             protected function setStoreValidationRules(): array
@@ -85,13 +86,135 @@ class BrickablesTest extends BrickableTestCase
     }
 
     /** @test */
+    public function it_only_displays_once_css_resources()
+    {
+        view()->addNamespace('laravel-brickables', 'tests/views');
+        $brickableOne = new class extends Brickable {
+            public function setBrickViewPath(): string
+            {
+                return 'laravel-brickables::brick-test';
+            }
+
+            protected function setCssResourcePath(): string
+            {
+                return 'my/test/css/path/brickable-one.css';
+            }
+
+            protected function setStoreValidationRules(): array
+            {
+                return [];
+            }
+
+            protected function setUpdateValidationRules(): array
+            {
+                return [];
+            }
+        };
+        $brickableTwo = new class extends Brickable {
+            public function setBrickViewPath(): string
+            {
+                return 'laravel-brickables::brick-test';
+            }
+
+            protected function setCssResourcePath(): string
+            {
+                return 'my/test/css/path/brickable-two.css';
+            }
+
+            protected function setStoreValidationRules(): array
+            {
+                return [];
+            }
+
+            protected function setUpdateValidationRules(): array
+            {
+                return [];
+            }
+        };
+        config()->set('brickables.registered', [get_class($brickableOne), get_class($brickableTwo)]);
+        $page = factory(Page::class)->create();
+        $page->addBricks([
+            [get_class($brickableOne), ['custom' => 'first-brickable-one']],
+            [get_class($brickableTwo), ['custom' => 'first-brickable-two']],
+            [get_class($brickableOne), ['custom' => 'second-brickable-one']],
+            [get_class($brickableTwo), ['custom' => 'second-brickable-two']],
+        ]);
+        $html = view('laravel-brickables::page-test', compact('page', 'brickableOne', 'brickableTwo'))->toHtml();
+        $this->assertStringNotContainsString($html, 'javascript');
+        $headHtmlContent = Str::between($html, '<head>', '</head');
+        $this->assertEquals(1, substr_count($headHtmlContent, 'my/test/css/path/brickable-one.css'));
+        $this->assertEquals(1, substr_count($headHtmlContent, 'my/test/css/path/brickable-two.css'));
+    }
+
+    /** @test */
+    public function it_only_displays_once_js_resources()
+    {
+        view()->addNamespace('laravel-brickables', 'tests/views');
+        $brickableOne = new class extends Brickable {
+            public function setBrickViewPath(): string
+            {
+                return 'laravel-brickables::brick-test';
+            }
+
+            protected function setJsResourcePath(): string
+            {
+                return 'my/test/js/path/brickable-one.js';
+            }
+
+            protected function setStoreValidationRules(): array
+            {
+                return [];
+            }
+
+            protected function setUpdateValidationRules(): array
+            {
+                return [];
+            }
+        };
+        $brickableTwo = new class extends Brickable {
+            public function setBrickViewPath(): string
+            {
+                return 'laravel-brickables::brick-test';
+            }
+
+            protected function setJsResourcePath(): string
+            {
+                return 'my/test/js/path/brickable-two.js';
+            }
+
+            protected function setStoreValidationRules(): array
+            {
+                return [];
+            }
+
+            protected function setUpdateValidationRules(): array
+            {
+                return [];
+            }
+        };
+        config()->set('brickables.registered', [get_class($brickableOne), get_class($brickableTwo)]);
+        $page = factory(Page::class)->create();
+        $page->addBricks([
+            [get_class($brickableOne), ['custom' => 'first-brickable-one']],
+            [get_class($brickableTwo), ['custom' => 'first-brickable-two']],
+            [get_class($brickableOne), ['custom' => 'second-brickable-one']],
+            [get_class($brickableTwo), ['custom' => 'second-brickable-two']],
+        ]);
+        $html = view('laravel-brickables::page-test', compact('page', 'brickableOne', 'brickableTwo'))->toHtml();
+        $bodyHtmlContent = Str::between($html, '<body>', '</body');
+        $this->assertStringNotContainsString($html, 'css');
+        $this->assertEquals(1, substr_count($bodyHtmlContent, 'my/test/js/path/brickable-one.js'));
+        $this->assertEquals(1, substr_count($bodyHtmlContent, 'my/test/js/path/brickable-two.js'));
+    }
+
+    /** @test */
     public function it_can_display_model_bricks_admin_panel_html()
     {
         view()->addNamespace('laravel-brickables', 'tests/views');
-        $brickable = new Class extends Brickable {
+        $brickable = new class extends Brickable {
             public function setBrickViewPath(): string
             {
-                return 'laravel-brickables::test-brick';
+                return 'laravel-brickables::brick-test';
             }
 
             protected function setStoreValidationRules(): array
@@ -117,7 +240,7 @@ class BrickablesTest extends BrickableTestCase
     /** @test */
     public function it_can_cast_bricks_to_their_brickable_related_brick_model()
     {
-        $brickableOne = new Class extends Brickable {
+        $brickableOne = new class extends Brickable {
             protected function setStoreValidationRules(): array
             {
                 return [];
@@ -128,7 +251,7 @@ class BrickablesTest extends BrickableTestCase
                 return [];
             }
         };
-        $brickableTwo = new Class extends Brickable {
+        $brickableTwo = new class extends Brickable {
             protected function setBrickModelClass(): string
             {
                 return BrickModel::class;
@@ -165,7 +288,7 @@ class BrickablesTest extends BrickableTestCase
     /** @test */
     public function it_can_cast_bricks_and_return_them_in_correct_order()
     {
-        $brickableOne = new Class extends Brickable {
+        $brickableOne = new class extends Brickable {
             protected function setStoreValidationRules(): array
             {
                 return [];
@@ -176,7 +299,7 @@ class BrickablesTest extends BrickableTestCase
                 return [];
             }
         };
-        $brickableTwo = new Class extends Brickable {
+        $brickableTwo = new class extends Brickable {
             protected function setBrickModelClass(): string
             {
                 return BrickModel::class;
@@ -205,7 +328,7 @@ class BrickablesTest extends BrickableTestCase
     /** @test */
     public function it_can_cast_brick_to_its_brickable_related_brick_model()
     {
-        $brickable = new Class extends Brickable {
+        $brickable = new class extends Brickable {
             protected function setBrickModelClass(): string
             {
                 return BrickModel::class;
@@ -245,7 +368,7 @@ class BrickablesTest extends BrickableTestCase
     /** @test */
     public function it_can_get_model_from_edit_request()
     {
-        $brickable = new Class extends Brickable {
+        $brickable = new class extends Brickable {
             protected function setStoreValidationRules(): array
             {
                 return [];