+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.0.0](https://github.com/Okipa/laravel-brickables/compare/1.1.0...2.0.0)
+
+2020-03-16
+
+* Dropped support for PHP versions under 7.4.
+* Dropped support for Laravel versions under 6.0.
+* Added javascript and css smart management with the `@brickables` blade directive.
+
 ## [1.1.0](https://github.com/Okipa/laravel-brickables/compare/1.0.2...1.1.0)
 
 2020-03-16